% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ders_report.R
\name{ders_stats_report}
\alias{ders_stats_report}
\title{DERS量表统计报告}
\usage{
ders_stats_report(data, prefix = "Cders")
}
\description{
生成DERS量表的总体信效度和分量表描述性统计
}
\details{
Generate DERS Scale Statistics Report
}
