% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_network_analysis.R
\name{dynamic_network_analysis}
\alias{dynamic_network_analysis}
\title{Dynamic Panel Network Analysis for Longitudinal Data}
\usage{
dynamic_network_analysis(
  data,
  id_col = "src_subject_id",
  time_col = "time",
  vars = NULL,
  groups = NULL,
  color = zcolor,
  threshold = 0.101,
  plot_colors = list(pos = c("#2376b7", "#134857"), neg = c("#d2568c", "#62102e")),
  output_dir = getwd(),
  file_prefix = "dynamicNetwork_dlvm1"
)
}
\arguments{
\item{data}{数据框，包含id_col、time_co和测量变量们}

\item{id_col}{ID列名}

\item{time_col}{时间列名}

\item{vars}{要分析的变量名向量}

\item{groups}{分组列表，是对 vars 的分组，用以上色}

\item{threshold}{网络阈值，三网络通用}

\item{plot_colors}{绘图连边颜色的简便设置}

\item{output_dir}{输出PDF的目录路径}

\item{file_prefix}{输出文件名前缀}
}
\value{
返回动态网络分析结果列表
}
\description{
Dynamic Panel Network Analysis for Longitudinal Data
}
