% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_network_analysis.R
\name{exploratory_network_analysis}
\alias{exploratory_network_analysis}
\title{Dynamic Panel Network Analysis for Longitudinal Data}
\usage{
exploratory_network_analysis(
  data,
  max_dimensions = 6,
  plot_output = TRUE,
  output_path = NULL,
  seed = 234
)
}
\arguments{
\item{data}{输入数据框，列为项目，行为被试
ena}

\item{seed}{随机数种子}

\item{output_dir}{图形输出路径}
}
\value{
返回探索络分析结果列表bootstrap_ega, stability, dimension_selection
}
\description{
Dynamic Panel Network Analysis for Longitudinal Data
}
