% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zdescribe.R
\name{print.quick_analysis}
\alias{print.quick_analysis}
\title{Print method for quick_analysis objects}
\usage{
\method{print}{quick_analysis}(x, ...)
}
\arguments{
\item{x}{Object of class quick_analysis}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for quick_analysis objects
}
