% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ders.R
\name{reverse_ders}
\alias{reverse_ders}
\title{DERS量表反向计分处理}
\usage{
reverse_ders(
  data,
  prefix = c("cders_", "pders_"),
  reverse_items = c(1, 2, 6, 7, 8, 10, 17, 20, 22, 24, 34),
  subscales = TRUE
)
}
\arguments{
\item{data}{数据框，包含DERS量表数据}

\item{prefix}{题目前缀，可以是"cders_"或"pders_"}

\item{reverse_items}{需要反向计分的题号向量}

\item{subscales}{是否计算分量表得分}
}
\value{
处理后的数据列表，包含处理后的数据和量表信息
}
\description{
处理情绪调节困难量表(DERS)的反向计分项目
}
\details{
Handle Reverse Scoring for DERS Scale
}
