% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_pd.R
\name{reverse_pd}
\alias{reverse_pd}
\title{人格倾向量表计分处理}
\usage{
reverse_pd(
  data,
  rater = c("child", "parent"),
  time_col = "eventname",
  subscales = TRUE
)
}
\arguments{
\item{data}{数据框}

\item{rater}{评分者类型，"child"或"parent"}

\item{time_col}{时间列名，默认为"eventname"}

\item{subscales}{是否计算分量表得分，默认TRUE}
}
\value{
处理后的数据列表
}
\description{
处理YPI、TFQ-20、CADS、ICU和RPAQ的反向计分和维度分计算
}
\details{
Handle Scoring for Personality Disposition Scales
}
