% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_cf.R
\name{risk_cf}
\alias{risk_cf}
\title{枪支风险评估分数计算}
\usage{
risk_cf(data, prefix = "pfire_")
}
\arguments{
\item{data}{数据框}

\item{prefix}{题目前缀，默认为"pfire_"}
}
\value{
包含风险评分的数据框
}
\description{
基于携带枪支历史计算风险评分
}
\details{
Calculate Firearm Risk Score
}
