% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_apq.R
\name{score_apq}
\alias{score_apq}
\title{APQ量表计分处理}
\usage{
score_apq(
  data,
  rater = c("child", "parent"),
  time_col = "eventname",
  subscales = TRUE
)
}
\arguments{
\item{data}{数据框}

\item{rater}{评分者类型，"child"或"parent"}

\item{time_col}{时间列名，默认为"eventname"}

\item{subscales}{是否计算分量表得分，默认TRUE}
}
\value{
处理后的数据列表
}
\description{
处理Alabama Parenting Questionnaire (APQ)的维度分计算
}
\details{
Handle Scoring for Alabama Parenting Questionnaire
}
