% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_nbh.R
\name{score_nbh}
\alias{score_nbh}
\title{邻里感知量表计分处理}
\usage{
score_nbh(data, prefix = "pneigh_", time_col = "eventname", subscales = TRUE)
}
\arguments{
\item{data}{数据框}

\item{prefix}{题目前缀，默认为"socialdev_pneigh_"}

\item{time_col}{时间列名，默认为"eventname"}

\item{subscales}{是否计算分量表得分，默认TRUE}
}
\value{
处理后的数据列表
}
\description{
处理邻里感知量表的反向计分和维度分计算
}
\details{
Handle Scoring for Neighborhood Perception Scale
}
\examples{
\dontrun{
result <- score_nbh(data)
}
}
