% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pb.R
\name{score_pb}
\alias{score_pb}
\title{应对行为维度计分处理}
\usage{
score_pb(data, prefix = "cpb_")
}
\arguments{
\item{data}{数据框}

\item{prefix}{题目前缀，默认为"cpb_"}
}
\value{
处理后的数据列表，包含原始数据和维度得分
}
\description{
处理应对行为的四个维度得分计算：积极学校、积极社交、消极学校、消极违法
}
\details{
Handle Scoring for Coping Behavior Dimensions
}
