% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pd.R
\name{score_pd}
\alias{score_pd}
\title{Simple PD Score Calculator}
\usage{
score_pd(data, rater)
}
\arguments{
\item{data}{数据框}

\item{rater}{"child" 或 "parent"}
}
\value{
处理后的数据
}
\description{
Simple PD Score Calculator
}
