% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_rd.R
\name{score_rd}
\alias{score_rd}
\title{自报违法行为量表计分处理}
\usage{
score_rd(data, prefix = "prd_", na_values = c(98, 99))
}
\arguments{
\item{data}{数据框}

\item{prefix}{题目前缀，默认为"prd_"}

\item{na_values}{需要被转换为NA的值向量，默认为c(98, 99)}
}
\value{
处理后的数据列表，包含清理后的数据和维度得分
}
\description{
处理SRD量表的数据清理和维度得分计算
}
\details{
Handle Scoring for Self-Reported Delinquency Scale
}
